ndt7-upload.js: send as aggressively as ndt5
It seems the ndt7 sender is more aggressive of the ndt5 one in
that the code loops refilling the buffer.

We already had issues with Safari with that pattern. We can see
that the ndt5 pattern is faster with Edge. So, a good question is
whether we can maybe try and use the same pattern of ndt5?